+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next
+  - Allow `version` be a numbers
+
 ## Changes in 0.23.0
   - Add support for custom decoders to allow for alternate syntax (e.g. Dhall)
   - `generated-exposed-modules` and `generated-other-modules`, for populating